♻️  consolidate duplicate code
